Split listeners and command
